Add more continuation stuff
